+++ packages/components-react/src/pages/index.tsx
@@ -141,8 +141,10 @@ import { ModalWithoutDismissButtonPage } from './generated/ModalWithoutDismissBu
 import { ModalWithoutHeadingPage } from './generated/ModalWithoutHeading';
 import { ModelSignaturePage } from './generated/ModelSignature';
 import { MultiSelectPage } from './generated/MultiSelect';
+import { MultiSelectCompactPage } from './generated/MultiSelectCompact';
 import { MultiSelectOpenedBottomPage } from './generated/MultiSelectOpenedBottom';
 import { MultiSelectOpenedFilterNoResultsPage } from './generated/MultiSelectOpenedFilterNoResults';
+import { MultiSelectOpenedFilterNoResultsCompactPage } from './generated/MultiSelectOpenedFilterNoResultsCompact';
 import { MultiSelectOpenedMaxHeightPage } from './generated/MultiSelectOpenedMaxHeight';
 import { MultiSelectOpenedMinHeightPage } from './generated/MultiSelectOpenedMinHeight';
 import { MultiSelectOpenedScrollContextPage } from './generated/MultiSelectOpenedScrollContext';
@@ -152,6 +154,11 @@ import { MultiSelectOpenedWithDisabledHighlightedSelectedPage } from './generate
 import { MultiSelectOpenedWithMultipleSelectedOptionsPage } from './generated/MultiSelectOpenedWithMultipleSelectedOptions';
 import { MultiSelectOpenedWithOptgroupsPage } from './generated/MultiSelectOpenedWithOptgroups';
 import { MultiSelectOpenedWithOptgroupsDisabledPage } from './generated/MultiSelectOpenedWithOptgroupsDisabled';
+import { MultiSelectScaling_0Page } from './generated/MultiSelectScaling_0';
+import { MultiSelectScaling_0_5Page } from './generated/MultiSelectScaling_0_5';
+import { MultiSelectScaling_0_8Page } from './generated/MultiSelectScaling_0_8';
+import { MultiSelectScaling_1Page } from './generated/MultiSelectScaling_1';
+import { MultiSelectScaling_2Page } from './generated/MultiSelectScaling_2';
 import { PaginationPage } from './generated/Pagination';
 import { PinCodePage } from './generated/PinCode';
 import { PopoverPage } from './generated/Popover';
@@ -867,6 +874,11 @@ export const generatedRoutes: RouteType[] = [
     element: <MultiSelectPage />,
   },
   {
+    name: 'Multi Select Compact',
+    path: '/multi-select-compact',
+    element: <MultiSelectCompactPage />,
+  },
+  {
     name: 'Multi Select Opened Bottom',
     path: '/multi-select-opened-bottom',
     element: <MultiSelectOpenedBottomPage />,
@@ -875,6 +887,11 @@ export const generatedRoutes: RouteType[] = [
     name: 'Multi Select Opened Filter No Results',
     path: '/multi-select-opened-filter-no-results',
     element: <MultiSelectOpenedFilterNoResultsPage />,
+  },
+  {
+    name: 'Multi Select Opened Filter No Results Compact',
+    path: '/multi-select-opened-filter-no-results-compact',
+    element: <MultiSelectOpenedFilterNoResultsCompactPage />,
   },
   {
     name: 'Multi Select Opened Max Height',
@@ -920,6 +937,31 @@ export const generatedRoutes: RouteType[] = [
     name: 'Multi Select Opened With Optgroups Disabled',
     path: '/multi-select-opened-with-optgroups-disabled',
     element: <MultiSelectOpenedWithOptgroupsDisabledPage />,
+  },
+  {
+    name: 'Multi Select Scaling 0',
+    path: '/multi-select-scaling-0',
+    element: <MultiSelectScaling_0Page />,
+  },
+  {
+    name: 'Multi Select Scaling 0 5',
+    path: '/multi-select-scaling-0-5',
+    element: <MultiSelectScaling_0_5Page />,
+  },
+  {
+    name: 'Multi Select Scaling 0 8',
+    path: '/multi-select-scaling-0-8',
+    element: <MultiSelectScaling_0_8Page />,
+  },
+  {
+    name: 'Multi Select Scaling 1',
+    path: '/multi-select-scaling-1',
+    element: <MultiSelectScaling_1Page />,
+  },
+  {
+    name: 'Multi Select Scaling 2',
+    path: '/multi-select-scaling-2',
+    element: <MultiSelectScaling_2Page />,
   },
   {
     name: 'Pagination',