+++ packages/components/src/components.d.ts
@@ -2018,6 +2018,10 @@ export namespace Components {
      */
     interface PMultiSelect {
         /**
+          * Displays as compact version.
+         */
+        "compact"?: boolean;
+        /**
           * The description text.
          */
         "description"?: string;
@@ -6208,6 +6212,10 @@ declare namespace LocalJSX {
      * @controlled { "props": ["value"], "event": "update", "isInternallyMutated": true }
      */
     interface PMultiSelect {
+        /**
+          * Displays as compact version.
+         */
+        "compact"?: boolean;
         /**
           * The description text.
          */