+++ packages/components/src/utils/select/filter.ts
@@ -1,23 +1,25 @@
-import { forceUpdate } from '@stencil/core';
 import type {
   MultiSelectOptgroup,
   MultiSelectOption,
 } from '../../components/multi-select/multi-select/multi-select-utils';
 import type { SelectOptgroup, SelectOption } from '../../components/select/select/select-utils';
+import { setHighlightedSelectOption } from './keyboard-behavior';
 
 export const updateFilterResults = (
   options: (SelectOption | MultiSelectOption)[],
   optgroups: (SelectOptgroup | MultiSelectOptgroup)[],
   filterValue: string
-): { hasFilterResults: boolean } => {
+): { hasFilterResults: boolean; resetCurrentlyHighlightedOption: boolean } => {
   const value = filterValue.toLowerCase();
+  let resetCurrentlyHighlightedOption = false;
 
   for (const option of options) {
     const matches = option.textContent.toLowerCase().includes(value);
-    // Highlighted state is only kept if highlighted option matches the filter, otherwise reset
+    // Highlighted state is only kept if highlighted option matches the filter, otherwise reset.
+    // Otherwise, the option could still be selected with enter despite being invisible.
     if (option.highlighted && !matches) {
-      option.highlighted = false;
-      forceUpdate(option);
+      setHighlightedSelectOption(option, false);
+      resetCurrentlyHighlightedOption = true;
     }
     // Use display none to preserve hidden state
     option.style.display = matches ? 'block' : 'none';
@@ -30,5 +32,8 @@ export const updateFilterResults = (
     (optgroup as HTMLOptGroupElement).style.display = visibleOptions ? 'block' : 'none';
   }
 
-  return { hasFilterResults: options.some((option) => option.style.display !== 'none' && !option.hidden) };
+  return {
+    hasFilterResults: options.some((option) => option.style.display !== 'none' && !option.hidden),
+    resetCurrentlyHighlightedOption,
+  };
 };